Move base - legend tests into separate file
